Add tests for OrbitSshQt (#4570)
This is using the new simple SSH server on the CI to test our SSH
implementation.

Tests for the SFTP subsystem is still missing and the coverage is not
perfect yet (mainly the error cases are still missing). Tests for SFTP
are coming though. The error cases will be trickier.

+++ src/OrbitSshQt/include/OrbitSshQt/Tunnel.h
@@ -68,7 +68,7 @@ class Tunnel : public StateMachineHelper<Tunnel, details::TunnelState> {
   }
 
  signals:
-  void tunnelOpened(uint16_t listen_port);
+  void tunnelOpened(int listen_port);
   void started();
   void stopped();
   void errorOccurred(std::error_code);
